pyproject.toml
[project]
name = "FastAPI-Boilerplate"
version = "0.1.0"
description = "Basic FastAPI Boilerplate"
readme = "README.md"
authors = [{name = "Don C John"}]
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
    "resend>=2.9.0",
]

[project.scripts]
demo-app = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "black>=23.3.0",
]
